Divide request and save access token
This will divide request and save access token function.
