+++ scripts/github/Github.js
@@ -9,10 +9,7 @@ import { GITHUB_CONFIG } from "./config/config.js";
  * @param {Object} payload - The optional payload associated with the action.
  */
 const dispatch = async (action, payload) => {
-  if (action === Github.REQUEST_AND_SAVE_ACCESS_TOKEN) {
-    const accessToken = await requestAndSaveAccessToken(payload);
-    getAuthenticatedUserInfo(accessToken);
-  } else if (action === Github.GET_REPOSITORIES) {
+  if (action === Github.GET_REPOSITORIES) {
     return await getAuthenticatedUserRepositories(payload);
   } else if (action === Github.COMMIT) {
     const response = await getShaForExistingFile(payload);
@@ -109,36 +106,6 @@ const getAuthenticatedUserRepositories = async () => {
 };
 
 /**
- * Retrieves authenticated user information.
- *
- * @param {string} accessToken - The access token for authentication.
- * @throws {Error} If the access token is falsy or if the API request fails.
- */
-const getAuthenticatedUserInfo = async (accessToken) => {
-  if (Util.isEmpty(accessToken)) {
-    throw new Error(Github.ERROR[Github.INVALID_ACCESS_TOKEN]);
-  }
-
-  const url = `${GITHUB_CONFIG.API_BASE_URL}/user`;
-  const headers = {
-    Authorization: `Bearer ${accessToken}`,
-  };
-
-  const response = await Util.request(url, 'GET', headers, undefined);
-  if (!response.ok) {
-    throw new Error(Github.ERROR[Github.FETCH_API_FAILED]);
-  }
-
-  const json = await response.json();
-  const githubID = json.login;
-  if (Util.isEmpty(githubID)) {
-    throw new Error(Github.ERROR[Github.NOT_FOUND_GITHUB_ID]);
-  }
-
-  chrome.storage.local.set({ githubID: githubID });
-};
-
-/**
  * Retrieves the SHA and sanitized content for an existing file from the GitHub repository.
  *
  * @param {Object} payload - The payload containing information about the file.
@@ -183,53 +150,6 @@ const getShaForExistingFile = async (payload) => {
     sha,
     content: sanitizedContent,
   };
-};
-
-/**
- * Requests and saves an access token.
- * Upon successful retrieval, the access token is saved to the local storage.
- *
- * @param {Object} payload - The payload containing the authorization code.
- * @param {string} payload.code - The code for authorization.
- * @returns {string} - The access token authenticated user.
- * @throws {Error} If the payload object is invalid or if the code is missing or invalid.
- * @throws {Error} If the API request fails or if the access token is not found in the response.
- */
-const requestAndSaveAccessToken = async (payload) => {
-  if (Util.isEmpty(payload)) {
-    throw new Error(Github.ERROR[Github.INVALID_PAYLOAD]);
-  }
-
-  const { code } = payload;
-  if (Util.isEmpty(code)) {
-    throw new Error(Github.ERROR[Github.INVALID_CODE]);
-  }
-
-  const url = `${GITHUB_CONFIG.BASE_URL}/login/oauth/access_token`;
-
-  const data = new FormData();
-  data.append('client_id', GITHUB_CONFIG.CLIENT_ID);
-  data.append('client_secret', GITHUB_CONFIG.CLIENT_SECRET);
-  data.append('code', code);
-
-  const response = await Util.request(url, 'POST', undefined, data);
-  if (!response.ok) {
-    throw new Error(Github.ERROR[Github.FETCH_API_FAILED]);
-  }
-
-  const text = await response.text();
-  const matchResult = text.match(/access_token=([^&]*)/);
-
-  if (Util.isEmpty(matchResult) || matchResult.length < 2) {
-    throw new Error(Github.ERROR[Github.NOT_FOUND_ACCESS_TOKEN]);
-  }
-
-  const accessToken = matchResult[1];
-  chrome.storage.local.set({ githubAccessToken: accessToken });
-
-  Util.closeLatestTab();
-
-  return accessToken;
 };
 
 export { dispatch };
