+++ scripts/github/constants.js
@@ -1,14 +1,5 @@
-export const BASE_URL = 'https://github.com';
-export const API_BASE_URL = 'https://api.github.com';
-export const CLIENT_ID = '2cfd0f1fae095d5a1684';
-export const CLIENT_SECRET = 'd1b31e8c09708b856ecbe03d1f9e9223e472da89';
-export const REDIRECT_URL = 'https://github.com';
-export const SCOPES = ['repo'];
-
 export const COMMIT = 'commit';
 export const GET_REPOSITORIES = 'getRepositories';
-export const OPEN_OAUTH_PAGE = 'openOauthPage';
-export const REQUEST_AND_SAVE_ACCESS_TOKEN = 'requestAndSaveAccessToken';
 
 // 오류 코드 정의
 export const SUCCESS_CODE = 'S200';
