+++ scripts/github/constants/actions.js
@@ -1,3 +1,4 @@
 export const ACTIONS = {
   OPEN_OAUTH_PAGE: 'openOauthPage',
+  REQUEST_AND_SAVE_ACCESS_TOKEN: 'requestAndSaveAccessToken',
 };
