+++ scripts/github/handlers/oauthHandler.js
@@ -1,0 +1,11 @@
+const currentUri = window.location.href;
+const code = currentUri.match(/code=([^&]*)/);
+if (code) {
+  const payload = { code: code[1] };
+
+  chrome.runtime.sendMessage({
+    platform: currentUri.includes('state') ? 'notion' : 'github',
+    action: 'requestAndSaveAccessToken',
+    payload,
+  });
+}
