+++ scripts/github/main.js
@@ -4,6 +4,8 @@ import { authService } from "./services/authService.js";
 export async function dispatch(action, payload) {
   const actions = {
     [ACTIONS.OPEN_OAUTH_PAGE]: () => authService.openOauthPage(),
+    [ACTIONS.REQUEST_AND_SAVE_ACCESS_TOKEN]: async () =>
+      await authService.requestAndSaveAccessToken(payload),
   };
 
   const handler = actions[action];
